requirements.txt
moviepy==2.0.0.dev2
opencv-python==4.8.1.78
numpy==1.24.3
pydub==0.25.1
openai-whisper==20231117
tqdm==4.66.1
Pillow==10.0.1
pyinstaller==5.13.0